Testing A For Loop That Changes Color
Called two variables: canvas2 and context2. Created a new variable called imageData, which will get the information from the first canvas. Testing a for loop that will change the red and green parameters to 0 and blue parameters to 225. Then the image should be uploaded to second canvas. Will test with this commit. Resource: https://stackoverflow.com/questions/27071351/change-the-color-of-each-pixel-in-an-image-java